Add two new functions and rework `orgmdb-fill-movie-properties`
Now `orgmdb-fill-movie-properties` tries to detect type of the current
header (if it's a movie/series/episode etc.). When called with
prefix-arg it also sets the headerline based on the type. It also
tries to check if the header contains imdb-id, and searches accordingly.

These functions are added:
- orgmdb-fill-movie-properties-with-imdb-id :: Asks for imdb id to
fill the properties.

- orgmdb-fill-movie-properties-with-title :: Asks for title and year
to fill the properties.

Also fixes #2